Stem a dataframe containing a column with sentences

Implementation of the SnowballC stemmer. Note that punctuation and capital letters are removed when processing.

pr_stem_sentences(df, col, language = "french")

Arguments

df
the data.frame containing the text

col
the column with the text

language
the language of the text. Defaut is french. See SnowballC::getStemLanguages() function for a list of supported languages.

Value

a tibble

Examples

a <- proustr::laprisonniere[1:10,]
pr_stem_sentences(a, text)

#> # A tibble: 10 x 4
#>                                                                                text
#>  *                                                                            <chr>
#>  1 des le matin la têt encor tourn contr le mur et avant d'avoir vu au dessus de gr
#>  2 quand je pens mainten que mon ami était venu à notr retour de balbec habit à par
#>  3 quand albertin sav par françois que dan la nuit de ma chambr aux rideau encor fe
#>  4                       le douleur sont de foll et qui le écout est encor plus fou
#>  5 je l'aim trop pour ne pas joyeux sourir de son mauv goût musical cet chanson du 
#>  6                                         une chanson d'adieu sort de sourc troubl
#>  7 une nu pass elle éclips le soleil je voi s'éteindr et rentr dan une grisaill le 
#>  8 le cloison qui sépar nos deux cabinet de toilet celui d'albertin tout pareil éta
#>  9 d'autr fois je rest couch rêv auss longtemp que je le voul car on avait ordre de
#> 10 je son françois j'ouvr le figaro j'y cherch et constat que ne s'y trouv pas un a
#> # ... with 3 more variables: book <chr>, volume <chr>, year <dbl>